+++ bot.py
@@ -1,4 +1,4 @@
-__version__ = "3.3.0-dev3"
+__version__ = "3.3.0-dev5"
 
 import asyncio
 import logging
@@ -583,7 +583,9 @@ class ModmailBot(commands.Bot):
 
         return sent_emoji, blocked_emoji
 
-    async def _process_blocked(self, message: discord.Message) -> bool:
+    async def _process_blocked(
+        self, message: discord.Message
+    ) -> typing.Tuple[bool, str]:
         sent_emoji, blocked_emoji = await self.retrieve_emoji()
 
         if str(message.author.id) in self.blocked_whitelisted_users:
@@ -591,13 +593,7 @@ class ModmailBot(commands.Bot):
                 self.blocked_users.pop(str(message.author.id))
                 await self.config.update()
 
-            if sent_emoji != "disable":
-                try:
-                    await message.add_reaction(sent_emoji)
-                except (discord.HTTPException, discord.InvalidArgument):
-                    logger.warning("Failed to add sent_emoji.", exc_info=True)
-
-            return False
+            return False, sent_emoji
 
         now = datetime.utcnow()
 
@@ -714,19 +710,62 @@ class ModmailBot(commands.Bot):
             reaction = sent_emoji
 
         await self.config.update()
+        return str(message.author.id) in self.blocked_users, reaction
+
+    @staticmethod
+    async def add_reaction(msg, reaction):
         if reaction != "disable":
             try:
-                await message.add_reaction(reaction)
+                await msg.add_reaction(reaction)
             except (discord.HTTPException, discord.InvalidArgument):
                 logger.warning("Failed to add reaction %s.", reaction, exc_info=True)
-        return str(message.author.id) in self.blocked_users
 
     async def process_dm_modmail(self, message: discord.Message) -> None:
         """Processes messages sent to the bot."""
-        blocked = await self._process_blocked(message)
-        if not blocked:
-            thread = await self.threads.find_or_create(message.author)
-            await thread.send(message)
+        blocked, reaction = await self._process_blocked(message)
+        if blocked:
+            return await self.add_reaction(message, reaction)
+        thread = await self.threads.find(recipient=message.author)
+        if thread is None:
+            if self.config["dm_disabled"] >= 1:
+                embed = discord.Embed(
+                    title=self.config["disabled_new_thread_title"],
+                    color=self.error_color,
+                    description=self.config["disabled_new_thread_response"],
+                )
+                embed.set_footer(
+                    text=self.config["disabled_new_thread_footer"],
+                    icon_url=self.guild.icon_url,
+                )
+                logger.info(
+                    "A new thread was blocked from %s due to disabled Modmail.",
+                    message.author,
+                )
+                _, blocked_emoji = await self.retrieve_emoji()
+                await self.add_reaction(message, blocked_emoji)
+                return await message.channel.send(embed=embed)
+            thread = self.threads.create(message.author)
+        else:
+            if self.config["dm_disabled"] == 2:
+                embed = discord.Embed(
+                    title=self.config["disabled_current_thread_title"],
+                    color=self.error_color,
+                    description=self.config["disabled_current_thread_response"],
+                )
+                embed.set_footer(
+                    text=self.config["disabled_current_thread_footer"],
+                    icon_url=self.guild.icon_url,
+                )
+                logger.info(
+                    "A message was blocked from %s due to disabled Modmail.",
+                    message.author,
+                )
+                _, blocked_emoji = await self.retrieve_emoji()
+                await self.add_reaction(message, blocked_emoji)
+                return await message.channel.send(embed=embed)
+
+        await self.add_reaction(message, reaction)
+        await thread.send(message)
 
     async def get_contexts(self, message, *, cls=commands.Context):
         """
@@ -910,13 +949,13 @@ class ModmailBot(commands.Bot):
 
             thread = await self.threads.find(channel=channel)
             if thread is not None and thread.recipient:
-                if await self._process_blocked(
-                    SimpleNamespace(
-                        author=thread.recipient,
-                        channel=SimpleNamespace(send=_void),
-                        add_reaction=_void,
+                if (
+                    await self._process_blocked(
+                        SimpleNamespace(
+                            author=thread.recipient, channel=SimpleNamespace(send=_void)
+                        )
                     )
-                ):
+                )[0]:
                     return
                 await thread.recipient.trigger_typing()